# Vectorize array size computation in Zip function.