--- a/velox/functions/prestosql/benchmarks/CMakeLists.txt
+++ b/velox/functions/prestosql/benchmarks/CMakeLists.txt
@@ -69,3 +69,6 @@ target_link_libraries(velox_functions_prestosql_benchmarks_map_input
 
 add_executable(velox_functions_benchmarks_url URLBenchmark.cpp)
 target_link_libraries(velox_functions_benchmarks_url ${BENCHMARK_DEPENDENCIES})
+
+add_executable(velox_functions_benchmarks_zip ZipBenchMark.cpp)
+target_link_libraries(velox_functions_benchmarks_url ${BENCHMARK_DEPENDENCIES})