--- a/velox/functions/prestosql/benchmarks/CMakeLists.txt
+++ b/velox/functions/prestosql/benchmarks/CMakeLists.txt
@@ -71,4 +71,4 @@ add_executable(velox_functions_benchmarks_url URLBenchmark.cpp)
 target_link_libraries(velox_functions_benchmarks_url ${BENCHMARK_DEPENDENCIES})
 
 add_executable(velox_functions_benchmarks_zip ZipBenchMark.cpp)
-target_link_libraries(velox_functions_benchmarks_url ${BENCHMARK_DEPENDENCIES})
+target_link_libraries(velox_functions_benchmarks_zip ${BENCHMARK_DEPENDENCIES})