```
============================================================================
/Users/kpai/src/Velox/velox/functions/prestosql/benchmarks/ZipBenchMark.cpprelative  time/iter  iters/s
============================================================================
zip                                                          3.44ms   290.91
vectorized_loop                                               2.87s  348.11m
default_loop                                     175.04%      1.64s  609.34m
lambda_loop                                                178.91ms     5.59
range_loop                                       104.31%   171.53ms     5.83
============================================================================
```